Major progress on MK8D (AArch32) boot:
- Scheduler: fiber-based cooperative scheduling, preemption, condvar/arbiter fixes
- Kernel: per-process Memory bridges, thread context reset parity
- Services: audio event ownership, nvhost GPU/AS parity, buffer queue fixes
- JIT (rdynarmic): NEON instruction fixes (VMOV imm8, VEXT, VUZP, VRINT, VCVT,
  VNEG, VABS), VectorExtract PALIGNR fix, STREX livelock fix
- Game boots, creates 54 threads, initializes GPU (Maxwell3D), reaches rendering
  setup after ~18 min of CPU-bound initialization

+++ audio_core/src/renderer/memory/pool_mapper.rs
@@ -8,14 +8,14 @@ use common::alignment::is_4kb_aligned;
 use common::ResultCode;
 
 pub struct PoolMapper<'a> {
-    process_handle: Option<u32>,
+    process_handle: Option<usize>,
     pool_infos: Option<&'a mut [MemoryPoolInfo]>,
     pool_count: u64,
     force_map: bool,
 }
 
 impl<'a> PoolMapper<'a> {
-    pub fn new(process_handle: Option<u32>, force_map: bool) -> Self {
+    pub fn new(process_handle: Option<usize>, force_map: bool) -> Self {
         Self {
             process_handle,
             pool_infos: None,
@@ -25,7 +25,7 @@ impl<'a> PoolMapper<'a> {
     }
 
     pub fn with_pools(
-        process_handle: Option<u32>,
+        process_handle: Option<usize>,
         pool_infos: &'a mut [MemoryPoolInfo],
         pool_count: u32,
         force_map: bool,
@@ -140,7 +140,7 @@ impl<'a> PoolMapper<'a> {
         self.force_map
     }
 
-    pub fn get_process_handle(&self, pool: &MemoryPoolInfo) -> Option<u32> {
+    pub fn get_process_handle(&self, pool: &MemoryPoolInfo) -> Option<usize> {
         match pool.get_location() {
             PoolLocation::Cpu => self.process_handle,
             PoolLocation::Dsp => None,
